ytdl: Mess with the output template
> Add the date, resolution (height), and the vid-id because 2 videos had
the same title, even though the date should be enough to differentiate
them
> I'll probably tweak this further, maybe I'll remove the resolution or
the id, the time at the beginning is great I can sort the files by name
or do all sorts of crazy things, whoa! ???
> I'll add a simple if statement to check for a playlist, just search
for `playlist` in the URL and change the output template to that, I
don't know what to put first. The date or the playlist id